fixing something

+++ src/database/migrations/20200607084312_expenses.js
@@ -2,7 +2,7 @@ exports.up = function (knex) {
   return knex.schema.createTable("expenses", (table) => {
     table.increments("id").primary();
     table.string("name").notNullable();
-    table.decimal("value").notNullable();
+    table.bigInteger("value").notNullable();
   });
 };
